Dashboard claim fix (#233)
* Add breaks in switch, remove log

* Fix Dashboard Claim/Switch button
